feat: implement EntryPoint for Tss, format files

+++ src/libraries/BTCStyleMerkle.sol
@@ -11,10 +11,7 @@ library BTCStyleMerkle {
      * @dev compute double-sha256(txA || txB)
      *      (where txA, txB are 32 bytes; after concatenation, it becomes 64 bytes; first sha256, then sha256)
      */
-    function doubleSha256Pair(
-        bytes32 txA,
-        bytes32 txB
-    ) public pure returns (bytes32) {
+    function doubleSha256Pair(bytes32 txA, bytes32 txB) public pure returns (bytes32) {
         // concatenate and do sha256 once
         bytes memory first = abi.encodePacked(txA, txB);
         bytes32 hash1 = sha256(first);
@@ -30,9 +27,7 @@ library BTCStyleMerkle {
      * @dev do double-sha256 for arbitrary bytes
      *      (in some scenarios, you need to concatenate arbitrary data before doing double SHA256, not just pair)
      */
-    function doubleSha256Bytes(
-        bytes memory data
-    ) public pure returns (bytes32) {
+    function doubleSha256Bytes(bytes memory data) public pure returns (bytes32) {
         bytes32 first = sha256(data);
         return sha256(abi.encodePacked(first));
     }
@@ -43,9 +38,7 @@ library BTCStyleMerkle {
      *      - note, here we assume each leaf is already in the 32 bytes internal order used in Bitcoin/Dogecoin,
      *        if you get txid from block explorer, you need to reverse it in advance.
      */
-    function computeMerkleRoot(
-        bytes32[] memory leaves
-    ) public pure returns (bytes32) {
+    function computeMerkleRoot(bytes32[] memory leaves) public pure returns (bytes32) {
         if (leaves.length == 0) {
             return bytes32(0);
         }
@@ -57,23 +50,17 @@ library BTCStyleMerkle {
         bytes32[] memory currentLevel = leaves;
 
         while (currentLevel.length > 1) {
-            uint newLength = (currentLevel.length + 1) / 2; // calculate the size of the new array after merging (add 1 if odd)
+            uint256 newLength = (currentLevel.length + 1) / 2; // calculate the size of the new array after merging (add 1 if odd)
 
             bytes32[] memory nextLevel = new bytes32[](newLength);
 
-            uint j = 0;
-            for (uint i = 0; i < currentLevel.length; i += 2) {
+            uint256 j = 0;
+            for (uint256 i = 0; i < currentLevel.length; i += 2) {
                 if (i + 1 == currentLevel.length) {
                     // when odd number, duplicate the last one
-                    nextLevel[j] = doubleSha256Pair(
-                        currentLevel[i],
-                        currentLevel[i]
-                    );
+                    nextLevel[j] = doubleSha256Pair(currentLevel[i], currentLevel[i]);
                 } else {
-                    nextLevel[j] = doubleSha256Pair(
-                        currentLevel[i],
-                        currentLevel[i + 1]
-                    );
+                    nextLevel[j] = doubleSha256Pair(currentLevel[i], currentLevel[i + 1]);
                 }
                 j++;
             }
@@ -92,7 +79,7 @@ library BTCStyleMerkle {
     function reverseBytes32(bytes32 input) public pure returns (bytes32) {
         // convert to temporary bytes, reverse, then convert back to bytes32
         bytes memory buf = new bytes(32);
-        for (uint i = 0; i < 32; i++) {
+        for (uint256 i = 0; i < 32; i++) {
             buf[i] = input[31 - i];
         }
         return bytes32(buf);
@@ -104,13 +91,10 @@ library BTCStyleMerkle {
      * @return bytes4
      */
     function reverseBytes4(uint32 input) public pure returns (bytes4) {
-        return
-            bytes4(
-                ((input & 0xff000000) >> 24) |
-                    ((input & 0x00ff0000) >> 8) |
-                    ((input & 0x0000ff00) << 8) |
-                    ((input & 0x000000ff) << 24)
-            );
+        return bytes4(
+            ((input & 0xff000000) >> 24) | ((input & 0x00ff0000) >> 8) | ((input & 0x0000ff00) << 8)
+                | ((input & 0x000000ff) << 24)
+        );
     }
 
     /**
@@ -137,10 +121,11 @@ library BTCStyleMerkle {
      * @return proof the Merkle proof for the given index
      * @return root the Merkle root of the leaves array
      */
-    function generateMerkleProof(
-        bytes32[] memory leaves,
-        uint256 index
-    ) public pure returns (bytes32[] memory proof, bytes32 root) {
+    function generateMerkleProof(bytes32[] memory leaves, uint256 index)
+        public
+        pure
+        returns (bytes32[] memory proof, bytes32 root)
+    {
         require(leaves.length > 0, "No tx hashes");
         require(index < leaves.length, "Invalid index");
 
@@ -159,9 +144,7 @@ library BTCStyleMerkle {
             for (uint256 i = 0; i < levelLen; i += 2) {
                 // when right node is missing, duplicate left node
                 bytes32 left = currentLevel[i];
-                bytes32 right = (i + 1 < levelLen)
-                    ? currentLevel[i + 1]
-                    : currentLevel[i];
+                bytes32 right = (i + 1 < levelLen) ? currentLevel[i + 1] : currentLevel[i];
 
                 uint256 parentIndex = i / 2;
                 nextLevel[parentIndex] = doubleSha256Pair(left, right);
@@ -186,12 +169,11 @@ library BTCStyleMerkle {
         // proof array may be shorter than maxDepth, no need to trim (or use dynamic array)
     }
 
-    function verifyMerkleProof(
-        bytes32 root,
-        bytes32[] memory proof,
-        bytes32 leaf,
-        uint256 index
-    ) public pure returns (bool) {
+    function verifyMerkleProof(bytes32 root, bytes32[] memory proof, bytes32 leaf, uint256 index)
+        public
+        pure
+        returns (bool)
+    {
         bytes32 computedHash = leaf;
 
         for (uint256 i = 0; i < proof.length; i++) {
